Extract settings
