+++ internal/backends/sqlite/metadata/metadata.go
@@ -15,16 +15,6 @@
 // Package metadata provides access to databases and collections information.
 package metadata
 
-import (
-	"database/sql"
-	"database/sql/driver"
-	"encoding/json"
-
-	"golang.org/x/exp/slices"
-
-	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
-)
-
 // Collection will probably have a method for getting column name / SQLite path expression for the given document field
 // once we implement field extraction.
 // IDColumn probably should go away.
@@ -60,76 +50,3 @@ func (c *Collection) deepCopy() *Collection {
 		Settings:  c.Settings.deepCopy(),
 	}
 }
-
-// Settings represents collection settings.
-type Settings struct {
-	Indexes []IndexInfo `json:"indexes"`
-}
-
-// deepCopy returns a deep copy.
-func (s Settings) deepCopy() Settings {
-	indexes := make([]IndexInfo, len(s.Indexes))
-
-	for i, index := range s.Indexes {
-		indexes[i] = IndexInfo{
-			Name:   index.Name,
-			Key:    slices.Clone(index.Key),
-			Unique: index.Unique,
-		}
-	}
-
-	return Settings{
-		Indexes: indexes,
-	}
-}
-
-// Value implements driver.Valuer interface.
-func (s Settings) Value() (driver.Value, error) {
-	res, err := json.Marshal(s)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	return string(res), nil
-}
-
-// Scan implements sql.Scanner interface.
-func (s *Settings) Scan(src any) error {
-	var err error
-
-	switch src := src.(type) {
-	case nil:
-		*s = Settings{}
-	case []byte:
-		err = json.Unmarshal(src, s)
-	case string:
-		err = json.Unmarshal([]byte(src), s)
-	default:
-		panic("can't scan collection settings")
-	}
-
-	if err != nil {
-		return lazyerrors.Error(err)
-	}
-
-	return nil
-}
-
-// IndexInfo represents information about a single index.
-type IndexInfo struct {
-	Name   string         `json:"name"`
-	Key    []IndexKeyPair `json:"key"`
-	Unique bool           `json:"unique"`
-}
-
-// IndexKeyPair consists of a field name and a sort order that are part of the index.
-type IndexKeyPair struct {
-	Field      string `json:"field"`
-	Descending bool   `json:"descending"`
-}
-
-// check interfaces
-var (
-	_ driver.Valuer = Settings{}
-	_ sql.Scanner   = (*Settings)(nil)
-)
